services/catalog-crawler/requirements.txt
openpyxl>=3.1
psycopg2-binary>=2.9
playwright>=1.44
